WiP: Movement

+++ app/src/main/java/sh/nami/pong/models/NVector.java
@@ -2,15 +2,15 @@ package sh.nami.pong.models;
 
 public class NVector extends Vector {
 
-    public NVector(float x, float y) {
+    public NVector(int x, int y) {
         super(x, y);
         normalise();
     }
 
     private void normalise(){
-        float magnitude = (float)Math.sqrt(this.getX() * this.getX() + this.getY() * this.getY());
-        float unitX = this.getX() / magnitude;
-        float unitY = this.getY() / magnitude;
+        int magnitude = (int)Math.sqrt(this.getX() * this.getX() + this.getY() * this.getY());
+        int unitX = this.getX() / magnitude;
+        int unitY = this.getY() / magnitude;
         this.setX(unitX);
         this.setY(unitY);
     }
